+++ src/main/java/br/com/belderrama/financesapi/repository/MovimentoRepository.java
@@ -3,7 +3,7 @@ package br.com.belderrama.financesapi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import br.com.belderrama.financesapi.model.Movimento;
 
-public interface MovimentoRepository extends JpaRepository<Movimento, Integer> {
+public interface MovimentoRepository extends JpaRepository<Movimento, Long> {
